requirements.txt
requests>=2.24.0
pyecharts>=1.9.0
tqdm
pandas>=1.1.0
numpy>=1.16.5
transitions
tushare
python-docx>=0.8.11
transitions
matplotlib
seaborn
Deprecated>=1.2.12
scikit-learn>=0.24.1
dill
openpyxl
pyarrow
loguru>=0.6.0
click
pytest
tenacity>=8.1.0
requests-toolbelt>=0.10.1
python-binance==1.0.15